requirements.txt
PyQt5
pandas
numpy
requests
python-binance
ta
matplotlib
mplfinance
scikit-learn
nltk
newsapi-python
